requirements.txt
ollama
SpeechRecognition
librosa
soundfile
pyaudio
python-dotenv
BeautifulSoup4